requirements.txt
pymongo
tqdm
jaro-winkler
requests
betfairlightweight
